+++ packages/braintree-integration/src/braintree-accelerated-checkout/braintree-accelerated-checkout-payment-strategy.spec.ts
@@ -1,8 +1,8 @@
 import { getScriptLoader } from '@bigcommerce/script-loader';
+import { noop } from 'lodash';
 
 import {
     InvalidArgumentError,
-    MissingDataError,
     OrderFinalizationNotRequiredError,
     PaymentArgumentInvalidError,
     PaymentInitializeOptions,
@@ -13,45 +13,57 @@ import {
     getShippingAddress,
     PaymentIntegrationServiceMock,
 } from '@bigcommerce/checkout-sdk/payment-integrations-test-utils';
+import { BrowserStorage } from '@bigcommerce/checkout-sdk/storage';
 
 import { BraintreeConnect } from '../braintree';
 import BraintreeIntegrationService from '../braintree-integration-service';
 import BraintreeScriptLoader from '../braintree-script-loader';
-import {
-    getBraintreeAcceleratedCheckoutPaymentMethod,
-    getConnectMock,
-} from '../mocks/braintree.mock';
+import { getConnectMock } from '../mocks/braintree.mock';
 
 import BraintreeAcceleratedCheckoutPaymentStrategy from './braintree-accelerated-checkout-payment-strategy';
+import BraintreeAcceleratedCheckoutUtils from './braintree-accelerated-checkout-utils';
 
 describe('BraintreeAcceleratedCheckoutPaymentStrategy', () => {
+    let braintreeAcceleratedCheckoutUtils: BraintreeAcceleratedCheckoutUtils;
     let braintreeConnectMock: BraintreeConnect;
     let braintreeIntegrationService: BraintreeIntegrationService;
     let braintreeScriptLoader: BraintreeScriptLoader;
+    let browserStorage: BrowserStorage;
     let paymentIntegrationService: PaymentIntegrationService;
     let strategy: BraintreeAcceleratedCheckoutPaymentStrategy;
 
-    const container = '#braintree-cc-component-container';
     const methodId = 'braintreeacceleratedcheckout';
-    const initializationOptions = {
-        methodId,
-        braintreeacceleratedcheckout: {
-            container,
-        },
-    };
+    const deviceSessionId = 'device_session_id_mock';
+    const instrumentId = 'asd123';
 
-    const paymentMethod = getBraintreeAcceleratedCheckoutPaymentMethod();
     const billingAddress = getBillingAddress();
     const shippingAddress = getShippingAddress();
 
-    const execureOptions = {
+    const defaultInitializationOptions = {
+        methodId,
+        braintreeacceleratedcheckout: {
+            onInit: jest.fn(),
+        },
+    };
+
+    const executeOptions = {
         payment: {
             methodId,
             paymentData: {},
         },
     };
 
+    const executeOptionsWithVaultedInstrument = {
+        payment: {
+            methodId,
+            paymentData: {
+                instrumentId,
+            },
+        },
+    };
+
     beforeEach(() => {
+        browserStorage = new BrowserStorage('braintree-accelerated-checkout-mock');
         braintreeConnectMock = getConnectMock();
 
         braintreeScriptLoader = new BraintreeScriptLoader(getScriptLoader(), window);
@@ -60,18 +72,20 @@ describe('BraintreeAcceleratedCheckoutPaymentStrategy', () => {
             window,
         );
         paymentIntegrationService = new PaymentIntegrationServiceMock();
+        braintreeAcceleratedCheckoutUtils = new BraintreeAcceleratedCheckoutUtils(
+            paymentIntegrationService,
+            braintreeIntegrationService,
+            browserStorage,
+        );
 
         strategy = new BraintreeAcceleratedCheckoutPaymentStrategy(
             paymentIntegrationService,
-            braintreeIntegrationService,
+            braintreeAcceleratedCheckoutUtils,
         );
 
         jest.spyOn(paymentIntegrationService, 'loadPaymentMethod');
         jest.spyOn(paymentIntegrationService, 'submitOrder');
         jest.spyOn(paymentIntegrationService, 'submitPayment');
-        jest.spyOn(paymentIntegrationService.getState(), 'getPaymentMethodOrThrow').mockReturnValue(
-            paymentMethod,
-        );
         jest.spyOn(
             paymentIntegrationService.getState(),
             'getBillingAddressOrThrow',
@@ -79,15 +93,30 @@ describe('BraintreeAcceleratedCheckoutPaymentStrategy', () => {
         jest.spyOn(paymentIntegrationService.getState(), 'getShippingAddress').mockReturnValue(
             shippingAddress,
         );
+        jest.spyOn(
+            paymentIntegrationService.getState(),
+            'getPaymentProviderCustomerOrThrow',
+        ).mockImplementation(jest.fn);
 
-        jest.spyOn(braintreeIntegrationService, 'initialize');
-        jest.spyOn(braintreeIntegrationService, 'getBraintreeConnect').mockImplementation(
-            () => braintreeConnectMock,
+        jest.spyOn(
+            braintreeAcceleratedCheckoutUtils,
+            'initializeBraintreeConnectOrThrow',
+        ).mockImplementation(jest.fn);
+        jest.spyOn(
+            braintreeAcceleratedCheckoutUtils,
+            'getBraintreeConnectOrThrow',
+        ).mockImplementation(() => braintreeConnectMock);
+        jest.spyOn(braintreeAcceleratedCheckoutUtils, 'getDeviceSessionId').mockImplementation(
+            () => deviceSessionId,
         );
         jest.spyOn(braintreeConnectMock, 'ConnectCardComponent').mockImplementation(() => ({
             tokenize: () => ({ nonce: 'nonce' }),
             render: jest.fn(),
         }));
+    });
+
+    afterEach(() => {
+        jest.clearAllMocks();
     });
 
     describe('#initialize()', () => {
@@ -125,53 +154,57 @@ describe('BraintreeAcceleratedCheckoutPaymentStrategy', () => {
                 expect(error).toBeInstanceOf(InvalidArgumentError);
             }
         });
+    });
+
+    describe('#renderBraintreeAXOComponent', () => {
+        let container: HTMLElement;
+        let callback: (containerId: string) => void = noop;
+
+        const initializationOptions = {
+            methodId,
+            braintreeacceleratedcheckout: {
+                onInit: (renderComponent: (containerId: string) => void) => {
+                    callback = renderComponent;
+                },
+            },
+        };
+
+        beforeEach(() => {
+            container = document.createElement('div');
+            container.id = 'pp-connect-container-id';
+            document.body.appendChild(container);
+        });
+
+        afterEach(() => {
+            document.body.removeChild(container);
+        });
+
+        it('throws an error if container id is not provided (or it has a falsy value) in callback call', async () => {
+            await strategy.initialize(initializationOptions);
+
+            try {
+                callback('');
+            } catch (error) {
+                expect(error).toBeInstanceOf(InvalidArgumentError);
+            }
+        });
 
         it('loads braintreeacceleratedcheckout payment method', async () => {
             await strategy.initialize(initializationOptions);
 
+            callback(container.id);
+
             expect(paymentIntegrationService.loadPaymentMethod).toHaveBeenCalledWith(methodId);
         });
 
-        it('throws an error if client token does not exist in payment method', async () => {
-            const state = paymentIntegrationService.getState();
-            const invalidPaymentMethod = {
-                ...paymentMethod,
-                clientToken: '',
-            };
-
-            jest.spyOn(state, 'getPaymentMethodOrThrow').mockReturnValue(invalidPaymentMethod);
-
-            try {
-                await strategy.initialize(initializationOptions);
-            } catch (error) {
-                expect(error).toBeInstanceOf(MissingDataError);
-            }
-        });
-
-        it('throws an error if initialization data does not exist in payment method', async () => {
-            const state = paymentIntegrationService.getState();
-            const invalidPaymentMethod = {
-                ...paymentMethod,
-                initializationData: null,
-            };
-
-            jest.spyOn(state, 'getPaymentMethodOrThrow').mockReturnValue(invalidPaymentMethod);
-
-            try {
-                await strategy.initialize(initializationOptions);
-            } catch (error) {
-                expect(error).toBeInstanceOf(MissingDataError);
-            }
-        });
-
-        it('loads braintree connect', async () => {
+        it('initializes braintree connect', async () => {
             await strategy.initialize(initializationOptions);
 
-            expect(braintreeIntegrationService.initialize).toHaveBeenCalledWith(
-                paymentMethod.clientToken,
-                paymentMethod.initializationData,
-            );
-            expect(braintreeIntegrationService.getBraintreeConnect).toHaveBeenCalled();
+            callback(container.id);
+
+            expect(
+                braintreeAcceleratedCheckoutUtils.initializeBraintreeConnectOrThrow,
+            ).toHaveBeenCalledWith(methodId);
         });
 
         it('renders braintree connect card component', async () => {
@@ -183,13 +216,17 @@ describe('BraintreeAcceleratedCheckoutPaymentStrategy', () => {
 
             await strategy.initialize(initializationOptions);
 
-            expect(renderMethodMock).toHaveBeenCalledWith(container);
+            if (container.id) {
+                callback(container.id);
+            }
+
+            expect(renderMethodMock).toHaveBeenCalledWith(container.id);
         });
     });
 
     describe('#execute()', () => {
         it('throws an error is payment is not provided', async () => {
-            await strategy.initialize(initializationOptions);
+            await strategy.initialize(defaultInitializationOptions);
 
             try {
                 await strategy.execute({});
@@ -199,8 +236,8 @@ describe('BraintreeAcceleratedCheckoutPaymentStrategy', () => {
         });
 
         it('submits payment and order with prepared data', async () => {
-            await strategy.initialize(initializationOptions);
-            await strategy.execute(execureOptions);
+            await strategy.initialize(defaultInitializationOptions);
+            await strategy.execute(executeOptions);
 
             expect(paymentIntegrationService.submitOrder).toHaveBeenCalledWith({}, undefined);
             expect(paymentIntegrationService.submitPayment).toHaveBeenCalledWith({
@@ -209,6 +246,72 @@ describe('BraintreeAcceleratedCheckoutPaymentStrategy', () => {
                     shouldSaveInstrument: false,
                     shouldSetAsDefaultInstrument: false,
                     nonce: 'nonce',
+                },
+            });
+        });
+
+        it('submits payment and order with stored instrument data', async () => {
+            await strategy.initialize(defaultInitializationOptions);
+            await strategy.execute(executeOptions);
+
+            expect(paymentIntegrationService.submitOrder).toHaveBeenCalledWith({}, undefined);
+            expect(paymentIntegrationService.submitPayment).toHaveBeenCalledWith({
+                methodId: 'braintreeacceleratedcheckout',
+                paymentData: {
+                    shouldSaveInstrument: false,
+                    shouldSetAsDefaultInstrument: false,
+                    nonce: 'nonce',
+                },
+            });
+        });
+    });
+
+    describe('#prepareVaultedInstrumentPaymentPayload()', () => {
+        it('prepares payment payload with BC vaulted instrument', async () => {
+            await strategy.initialize(defaultInitializationOptions);
+            await strategy.execute(executeOptionsWithVaultedInstrument);
+
+            expect(paymentIntegrationService.submitPayment).toHaveBeenCalledWith({
+                methodId: 'braintreeacceleratedcheckout',
+                paymentData: {
+                    deviceSessionId,
+                    instrumentId,
+                },
+            });
+        });
+
+        it('prepares payment payload with PayPal Connect vaulted instrument', async () => {
+            jest.spyOn(
+                paymentIntegrationService.getState(),
+                'getPaymentProviderCustomerOrThrow',
+            ).mockImplementation(() => ({
+                authenticationState: 'succeeded',
+                instruments: [
+                    {
+                        brand: 'visa',
+                        expiryMonth: '12',
+                        expiryYear: '33',
+                        iin: '411111',
+                        last4: '1111',
+                        type: 'card',
+                        bigpayToken: instrumentId,
+                        defaultInstrument: false,
+                        provider: methodId,
+                        trustedShippingAddress: false,
+                        method: methodId,
+                    },
+                ],
+            }));
+
+            await strategy.initialize(defaultInitializationOptions);
+            await strategy.execute(executeOptionsWithVaultedInstrument);
+
+            expect(paymentIntegrationService.submitPayment).toHaveBeenCalledWith({
+                methodId: 'braintreeacceleratedcheckout',
+                paymentData: {
+                    deviceSessionId,
+                    instrumentId,
+                    tokenType: 'paypal_connect',
                 },
             });
         });
@@ -223,8 +326,8 @@ describe('BraintreeAcceleratedCheckoutPaymentStrategy', () => {
                 render: jest.fn,
             }));
 
-            await strategy.initialize(initializationOptions);
-            await strategy.execute(execureOptions);
+            await strategy.initialize(defaultInitializationOptions);
+            await strategy.execute(executeOptions);
 
             expect(tokenizeMethodMock).toHaveBeenCalledWith({
                 billingAddress: {
@@ -246,13 +349,8 @@ describe('BraintreeAcceleratedCheckoutPaymentStrategy', () => {
     });
 
     describe('#deinitialize()', () => {
-        it('runs teardown process for braintree connect', async () => {
-            jest.spyOn(braintreeIntegrationService, 'teardown');
-
-            await strategy.initialize(initializationOptions);
-            await strategy.deinitialize();
-
-            expect(braintreeIntegrationService.teardown).toHaveBeenCalled();
+        it('deinitializes strategy', async () => {
+            await expect(strategy.deinitialize()).resolves.not.toThrow();
         });
     });