+++ packages/braintree-integration/src/braintree-accelerated-checkout/create-braintree-accelerated-checkout-customer-strategy.ts
@@ -4,12 +4,14 @@ import {
     CustomerStrategyFactory,
     toResolvableModule,
 } from '@bigcommerce/checkout-sdk/payment-integration-api';
+import { BrowserStorage } from '@bigcommerce/checkout-sdk/storage';
 
 import { BraintreeHostWindow } from '../braintree';
 import BraintreeIntegrationService from '../braintree-integration-service';
 import BraintreeScriptLoader from '../braintree-script-loader';
 
 import BraintreeAcceleratedCheckoutCustomerStrategy from './braintree-accelerated-checkout-customer-strategy';
+import BraintreeAcceleratedCheckoutUtils from './braintree-accelerated-checkout-utils';
 
 const createBraintreeAcceleratedCheckoutCustomerStrategy: CustomerStrategyFactory<
     BraintreeAcceleratedCheckoutCustomerStrategy
@@ -19,10 +21,16 @@ const createBraintreeAcceleratedCheckoutCustomerStrategy: CustomerStrategyFactor
         new BraintreeScriptLoader(getScriptLoader(), braintreeHostWindow),
         braintreeHostWindow,
     );
+    const browserStorage = new BrowserStorage('paypal-connect');
+    const braintreeAcceleratedCheckoutUtils = new BraintreeAcceleratedCheckoutUtils(
+        paymentIntegrationService,
+        braintreeIntegrationService,
+        browserStorage,
+    );
 
     return new BraintreeAcceleratedCheckoutCustomerStrategy(
         paymentIntegrationService,
-        braintreeIntegrationService,
+        braintreeAcceleratedCheckoutUtils,
     );
 };