+++ packages/core/src/payment-provider-customer/payment-provider-customer.ts
@@ -1,10 +1,10 @@
-import { AddressRequestBody } from '../address';
+import { CustomerAddress } from '../customer';
 import { CardInstrument } from '../payment/instrument/instrument';
 
 export type PaymentProviderCustomer = BraintreeAcceleratedCheckoutCustomer;
 
 export interface BraintreeAcceleratedCheckoutCustomer {
     authenticationState?: string;
-    addresses?: AddressRequestBody[];
+    addresses?: CustomerAddress[];
     instruments?: CardInstrument[];
 }